logic push method changed
